requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
kafka-python==2.0.1
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
